+++ internal/withdraw/withdraw.go
@@ -8,7 +8,7 @@ import (
 type Withdraw struct {
 	OrderNumber string    `json:"order,omitempty"`
 	UserId      int       `json:"-"`
-	Sum         int       `json:"sum,omitempty"`
+	Sum         float32   `json:"sum,omitempty"`
 	UploadData  time.Time `json:"processed_at,omitempty"`
 }
